All complete and working

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: gsiddiqu <gsiddiqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/17 13:53:51 by gsiddiqu          #+#    #+#             */
-/*   Updated: 2021/06/21 18:29:35 by gsiddiqu         ###   ########.fr       */
+/*   Updated: 2021/06/25 17:16:59 by gsiddiqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
