+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: gsiddiqu <gsiddiqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/17 13:56:45 by gsiddiqu          #+#    #+#             */
-/*   Updated: 2021/06/21 19:30:32 by gsiddiqu         ###   ########.fr       */
+/*   Updated: 2021/06/25 17:20:39 by gsiddiqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
