+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: gsiddiqu <gsiddiqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/17 13:54:15 by gsiddiqu          #+#    #+#             */
-/*   Updated: 2021/06/21 19:31:26 by gsiddiqu         ###   ########.fr       */
+/*   Updated: 2021/06/25 17:20:08 by gsiddiqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && (dst != NULL || src != NULL))
