+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: gsiddiqu <gsiddiqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/17 12:38:34 by gsiddiqu          #+#    #+#             */
-/*   Updated: 2021/06/21 20:07:47 by gsiddiqu         ###   ########.fr       */
+/*   Updated: 2021/06/25 17:28:13 by gsiddiqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	int		ind;
-	int		size;
-	char	*sub;
+	size_t		ind;
+	size_t		size;
+	char		*sub;
 
 	if (s == NULL)
 		return (NULL);
@@ -27,7 +27,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	if (!sub)
 		return (sub);
 	ind = 0;
-	while (ind < size && (start < ft_strlen((char *)s)))
+	while (ind < size && ((int)start < ft_strlen((char *)s)))
 	{
 		*(sub + ind) = *(s + start + ind);
 		ind++;
